[L] Implement publisher information api - ref #202
