+++ app/__init__.py
@@ -18,6 +18,7 @@ from app.utils import get_s3_cdn_prefix
 from app.auth.controllers import auth_blueprint
 from app.package.controllers import package_blueprint
 from app.site.controllers import site_blueprint
+from app.profile.controllers import profile_blueprint
 
 app_config = {
     "base": "app.config.BaseConfig",
@@ -45,6 +46,7 @@ def create_app():
     app.register_blueprint(package_blueprint)
     app.register_blueprint(auth_blueprint)
     app.register_blueprint(site_blueprint)
+    app.register_blueprint(profile_blueprint)
 
     s3 = boto3.client('s3',
                       region_name=app.config['AWS_REGION'],
