+++ tests/package/test_models.py
@@ -11,8 +11,8 @@ import unittest
 import json
 from app import create_app
 from app.database import db
-from app.package.models import BitStore, Package, User, \
-    Publisher, PublisherUser, UserRoleEnum, PackageStateEnum
+from app.package.models import BitStore, Package, PackageStateEnum
+from app.profile.models import User, Publisher, UserRoleEnum, PublisherUser
 
 
 class BitStoreTestCase(unittest.TestCase):
@@ -437,87 +437,12 @@ class PackageTestCase(unittest.TestCase):
 
         Package.create_or_update_version(self.publisher_one,
                                          self.package_two,
-                                            'tag_one')
+                                         'tag_one')
         tagged_data = Package.query.join(Publisher). \
             filter(Publisher.name == self.publisher_one,
                    Package.name == self.package_two,
                    Package.version == 'tag_one').one()
         self.assertEqual(latest_data.readme, tagged_data.readme)
-
-    def tearDown(self):
-        with self.app.app_context():
-            db.session.remove()
-            db.drop_all()
-
-
-class PublisherUserTestCase(unittest.TestCase):
-
-    def setUp(self):
-        self.app = create_app()
-        self.app.app_context().push()
-
-    def test_throw_error_if_role_is_invalid(self):
-        with self.app.test_request_context():
-            db.drop_all()
-            db.create_all()
-            publisher = Publisher(name='test_pub_id')
-            association = PublisherUser()
-            association.publisher = publisher
-            self.assertRaises(ValueError, association.role, "NOT_MEMBER")
-
-    def tearDown(self):
-        with self.app.app_context():
-            db.session.remove()
-            db.drop_all()
-
-
-class UserTestCase(unittest.TestCase):
-    def setUp(self):
-        self.app = create_app()
-        self.app.app_context().push()
-        with self.app.test_request_context():
-            db.drop_all()
-            db.create_all()
-
-            user = User(id=11,
-                        name='test_user_id',
-                        secret='supersecret',
-                        auth0_id="123|auth0")
-            publisher = Publisher(name='test_pub_id')
-            association = PublisherUser(role=UserRoleEnum.owner)
-            association.publisher = publisher
-            user.publishers.append(association)
-
-            db.session.add(user)
-            db.session.commit()
-
-    def test_serialize(self):
-        user = User.query.filter_by(name='test_user_id').one() \
-            .serialize
-        self.assertEqual('test_user_id', user['name'])
-
-    def test_user_role_on_publisher(self):
-        user = User.query.filter_by(name='test_user_id').one()
-        self.assertEqual(len(user.publishers), 1)
-        self.assertEqual(user.publishers[0].role, UserRoleEnum.owner)
-
-    def test_user_creation_from_outh0_response(self):
-        user_info = dict(email="test@test.com",
-                         username="test",
-                         user_id="124|auth0")
-        user = User.create_or_update_user_from_callback(user_info)
-        self.assertEqual(user.name, 'test')
-
-    def test_update_secret_if_it_is_supersecret(self):
-        user_info = dict(email="test@test.com",
-                         username="test",
-                         user_id="123|auth0")
-        user = User.create_or_update_user_from_callback(user_info)
-        self.assertNotEqual('supersecret', user.secret)
-
-    def test_get_userinfo_by_id(self):
-        self.assertEqual(User.get_userinfo_by_id(11).name, 'test_user_id')
-        self.assertIsNone(User.get_userinfo_by_id(2))
 
     def tearDown(self):
         with self.app.app_context():
